[FIX] resource_activity: cancelled registration cancels allocation
Before: booking and cancellation on registration is done through
resource.available. There is no direct link between the actual allocation
and the resource available. When cancelling a registration,
the linked resource available are cancelled, the resource linked
to the resource available for that time is searched and cancelled.
The resource might have changed on the allocation but not on the available.
In that case, the wrong allocation or no allocation will be cancelled.

After: there is a direct link between the allocation and the resource available.
when cancelling a registration, the allocation is cancelled and all the
available are cleaned up.

complete story :
https://gestion.coopiteasy.be/web#id=10293&action=475&active_id=223&model=project.task&view_type=form&menu_id=536